Changed enemy collision to send player back instead
There was a bug where the chest would fall through the platforms on stage 1 so I turned off the gravity for the chest to fix it on that stage. Also, there is a bug where if you get a letter and then lose the letter and then pickup a different instance of the lost letter (for example more than one P can spawn on the map), so lets say you pick up a P at the beginning of the stage and then lose it, but you pick up a P at a different part of the stage, it will update the letterboard but it will not let you win the game.